\name{CSREC-package}
\alias{CSREC-package}
\alias{CSREC}
\docType{package}
\title{
\packageTitle{CSREC}
}
\description{
\packageDescription{CSREC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CSREC}
\packageIndices{CSREC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CSREC}

Maintainer: \packageMaintainer{CSREC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
